# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "My API Documentation",
    "version": "1.0.0",
    "description": "API documentation for my Next.js app"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    },
    {
      "url": "https://book-teller-batrinutsilvius-projects.vercel.app",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your JWT token from the login endpoint"
      }
    }
  },
  "paths": {
...
    "/api/playlists/{playlist_id}/languages/{language_id}": {
      "get": {
        "summary": "Get a playlist by ID",
        "tags": [
          "Playlists"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "playlist_id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Playlist ID"
          },
          {
            "in": "path",
            "name": "language_id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Language ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Playlist with books and their translations for the given language",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer"
                    },
                    "profile_id": {
                      "type": "integer"
                    },
                    "name": {
                      "type": "string"
                    },
                    "created_at": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "playlistBooks": {
                      "type": "array",
                      "description": "Sorted by order ascending",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "integer"
                          },
                          "playlist_id": {
                            "type": "integer"
                          },
                          "book_id": {
                            "type": "integer"
                          },
                          "order": {
                            "type": "integer"
                          },
                          "created_at": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "book": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "integer"
                              },
                              "photo_url": {
                                "type": "string",
                                "nullable": true
                              },
                              "bookTranslations": {
                                "type": "array",
                                "description": "Contains one translation for the requested language",
                                "items": {
                                  "type": "object",
                                  "properties": {
                                    "id": {
                                      "type": "integer"
                                    },
                                    "title": {
                                      "type": "string"
                                    }
                                  }
                                }
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - not your playlist"
          },
          "404": {
            "description": "Playlist or language not found"
          }
        }
      }
    },
...
    "/api/playlists/{playlist_id}/books": {
      "post": {
        "summary": "Add book to playlist",
        "tags": [
          "Playlists"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "playlist_id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Playlist ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "book_id"
                ],
                "properties": {
                  "book_id": {
                    "type": "integer",
                    "description": "ID of the book to add",
                    "example": 5
                  },
                  "position": {
                    "type": "integer",
                    "description": "Position in playlist (optional, defaults to end)",
                    "example": 2
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Book added to playlist successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer"
                    },
                    "playlist_id": {
                      "type": "integer"
                    },
                    "book_id": {
                      "type": "integer"
                    },
                    "order": {
                      "type": "integer"
                    },
                    "book": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer"
                        },
                        "photo_url": {
                          "type": "string",
                          "nullable": true
                        },
                        "bookTranslations": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "integer"
                              },
                              "title": {
                                "type": "string"
                              },
                              "language_id": {
                                "type": "integer"
                              },
                              "description": {
                                "type": "string",
                                "nullable": true
                              },
                              "language": {
                                "type": "object",
                                "properties": {
                                  "name": {
                                    "type": "string"
                                  },
                                  "country_code": {
                                    "type": "string"
                                  }
                                }
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - validation error"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - playlist doesn't belong to user"
          },
          "404": {
            "description": "Playlist or book not found"
          },
          "409": {
            "description": "Book already in playlist"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
...
  },
...
}